retail_app/backend/requirements.txt
anyio==4.8.0
azure-core==1.32.0
azure-identity==1.21.0
backoff==2.2.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deltalake==0.25.4
flask==3.1.0
flask-caching==2.3.1
flask-cors==5.0.1
gql==3.5.1
graphql-core==3.2.3
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
msal==1.31.1
msal-extensions==1.3.1
multidict==6.1.0
propcache==0.3.0
pyarrow==18.1.0
pycparser==2.22
pyjwt==2.10.1
pyodbc==5.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
yarl==1.18.3
